+++ src/main/config.ts
@@ -8,6 +8,7 @@ const __dirname = dirname(fileURLToPath(import.meta.url))
 
 export const APP_CONFIG = {
   name: "Daily",
+  author: "Scheron",
   protocol: "daily",
   filesProtocol: "daily://file",
   iCloudPath: `${process.env.HOME}/Library/Mobile Documents/com~apple~CloudDocs`,